requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
drf-spectacular>=0.27
requests>=2.31
python-dotenv>=1.0
PyMySQL>=1.1
